feat: move tag.createNode to static nodeClass.from
Re: https://github.com/eemeli/yaml/issues/457
